Time: 39 ms (56.62%), Space: 105.7 MB (91.01%) - LeetHub

+++ 1966-frequency-of-the-most-frequent-element/1966-frequency-of-the-most-frequent-element.cpp
@@ -13,7 +13,7 @@ public:
                 total-=nums[i];
                 i++;
             }
-           res=max(res,(long long)j-i+1);
+           res=max(res,j-i+1);
            j++;
         }
         return res;
